requirements.txt
pandas
requests
fpdf
python-dotenv
